Protobuf support for more things that go by wire.
Before getting into replication I want everything to use protobuf.

It would be a shame to spend so much effort to get real live P2P working
only to have to be left with technical debt of a half baked
serialisation effort.

+++ query.go
@@ -56,7 +56,7 @@ type QueryRowJoin struct {
 
 type QuerySelect struct {
 	Where QueryWhere `json:",omitempty"`
-	Limit uint       `json:",omitempty"`
+	Limit uint64     `json:",omitempty"`
 }
 
 type QueryWhereOpCode uint16
